refactor project name
