+++ src/teamster/kippcamden/__init__.py
@@ -3,7 +3,6 @@ import pendulum
 from teamster.core.utils.classes import FiscalYear
 
 CODE_LOCATION = "kippcamden"
-GCS_PROJECT_NAME = "teamster-332318"
 
 LOCAL_TIMEZONE = pendulum.timezone(name="America/New_York")
 NOW = pendulum.now(tz=LOCAL_TIMEZONE)
