+++ tests/fivetran/test_fivetran_sensor.py
@@ -4,7 +4,7 @@ from dagster import EnvVar, build_sensor_context
 from dagster_fivetran import FivetranResource
 from dagster_gcp import BigQueryResource
 
-from teamster.kipptaf import GCS_PROJECT_NAME
+from teamster import GCS_PROJECT_NAME
 from teamster.kipptaf.fivetran import assets
 from teamster.kipptaf.fivetran.sensors import build_fivetran_sync_status_sensor
