+++ tests/google/bigquery/test_ops_bigquery.py
@@ -1,11 +1,11 @@
 from dagster import build_op_context
 from dagster_gcp import BigQueryResource
 
+from teamster import GCS_PROJECT_NAME
 from teamster.core.google.bigquery.ops import (
     BigQueryGetTableOpConfig,
     bigquery_get_table_op,
 )
-from teamster.kipptaf import GCS_PROJECT_NAME
 
 
 def test_op():
