+++ tests/schoolmint/test_schoolmint_ops.py
@@ -1,13 +1,13 @@
 from dagster import EnvVar, build_op_context
 from dagster_gcp import BigQueryResource
 
+from teamster import GCS_PROJECT_NAME
 from teamster.core.google.bigquery.ops import (
     BigQueryGetTableOpConfig,
     bigquery_get_table_op,
     bigquery_query_op,
 )
 from teamster.core.schoolmint.grow.resources import SchoolMintGrowResource
-from teamster.kipptaf import GCS_PROJECT_NAME
 from teamster.kipptaf.schoolmint.ops import (
     schoolmint_grow_school_update_op,
     schoolmint_grow_user_delete_op,
